move global state caching logic to expose global state
